Desafio02/requirements.txt
requests==2.32.3
behave==1.2.6
coverage==7.6.1
Appium-Python-Client==4.0.0